requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
openai
python-dotenv